requirements.txt
streamlit
urllib3
numpy
pandas
Pillow
requests
beautifulsoup4
